requirements.txt
requests
python-dotenv
tzdata
discord.py